@@ -1,40 +1,8 @@
 export const environment = {
   name: 'prod',
   production: true,
-  API_BASE_URL: 'https://mak-api.vercel.app/api/',
-  DOMAIN_URL: 'https://mak-ui.vercel.app',
-  bugsnag: '0d02ac2f63382263e24f75c0adad288e',
-  clevertap_id: '4R6-478-RR4Z',
-  facebook_id: '287403324698248',
-  facebook_pixel: '441308802710601',
-  gcaptcha: '6Lemuy0UAAAAAAJL2gRKu1BzpYtg5oI8QJCT89bI',
-  google_client_id: '154316787884-v1hbq46oa8g9vpfnt7m7s9b6i3cul0pc.apps.googleusercontent.com',
-  google_client_id_one_tap: '154316787884-v1hbq46oa8g9vpfnt7m7s9b6i3cul0pc.apps.googleusercontent.com',
-  gtm_id: 'GTM-MJRMLZ',
-  gtm_ids:<any> {
-    'mak': 'GTM-MJRMLZ',
-    'equalall': 'GTM-MJRMLZ',
-    'ketto': 'GTM-MJRMLZ'
-  },
-  invisible_captcha: '6LcYhV8UAAAAABM9Km9bMITOAJyPEJNPM3uYnuQL',
-  microsoft_clarity: 'nr5a4ti0hm',
-  puhser_key: 'e8bdd5e81bddad457212',
-  meilisearch: {
-    host: 'https://msearch.ketto.org',
-    key: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzZWFyY2hSdWxlcyI6eyJwcm9kX2Rpc2Vhc2VzIjp7fSwicHJvZF9ob3NwaXRhbHMiOnt9LCJmdW5kcmFpc2VyX3Byb2QiOnsiZmlsdGVyIjoicGFyZW50X2NhdXNlX2lkICE9IDE0OSJ9fSwiYXBpS2V5VWlkIjoiMjk0YjYzMTAtZjdiNC00MDdiLTg2MWYtMGQ0MTJmMzdkZjc4IiwiaWF0IjoxNjc2ODg3NzkzfQ.CXJdTbdG2JMQPQYnCuXsrv0oFlOjDs8hO3w0Rep4lUE',
-    index: 'fundraiser_prod',
-    disease: {
-      index: 'prod_diseases',
-      not_listed: 4562
-    },
-    hospital: {
-      index: 'prod_hospitals',
-      not_listed: 1734593,
-      not_admitted: 1734594,
-      no_hospitalisation_required: 1734595
-    }
-  },
-  truecaller: {
-    partnerKey: 'IYGXye7642585d0924b7086aff71119217401',
+  host: {
+    baseUrl: 'https://mak-ui.vercel.app',
+    apiBaseUrl: 'https://mak-api.vercel.app/api/'
   }
 };
